src/main/Test/aruco-app/server/requirements.txt
numpy==1.26.4   
opencv-python-headless==4.9.0.80
fastapi==0.111.0
uvicorn==0.29.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0   
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0

